AGVis backend restructure
